* ErrorMessageSubmits moved to submits sub-package * `ErrorMessageSubmitWithoutColumn` changed from `case class` to `class` to allow inheritance * some PR comments addressed
